Update HomePage and DisplayJSON component with Select feature
This update introduces a Select function which allows users to choose from a variety of options. It also includes adjustments to manage the display and layouts more efficiently, resulting in improved user interaction and overall user experience. Critical enhancements have been done to how data and key descriptions are handled.